+++ BlasterMaster/Engine/Renderer/Sprite.cc
@@ -1,12 +1,12 @@
 #include "Sprite.hh"
 #include "Renderer.hh"
 
-Sprite::Sprite(size_t ID, size_t top, size_t left, size_t bottom, size_t right, Texture* texture) 
+Sprite::Sprite(size_t ID, float top, float left, float bottom, float right, Texture* texture)
 {
-  m_Rect.top    = top;
-  m_Rect.left   = left;
-  m_Rect.bottom = bottom;
-  m_Rect.right  = right;
+  m_Rect.top    = (LONG)top;
+  m_Rect.left   = (LONG)left;
+  m_Rect.bottom = (LONG)bottom;
+  m_Rect.right  = (LONG)right;
   m_Texture     = texture;
 }
 
@@ -27,10 +27,10 @@ void Sprite::Render(float X, float Y)
 
 std::shared_ptr<SpriteBase> SpriteBase::s_Instance = std::make_shared<SpriteBase>();
 
-Sprite* SpriteBase::Add(size_t ID, size_t top, size_t left, size_t bottom, size_t right, Texture* texture)
+Sprite* SpriteBase::Add(size_t ID, float top, float left, float bottom, float right, Texture* texture)
 {
   m_Sprites[ID].reset(new Sprite(ID, top, left, bottom, right, texture));
-  return m_Sprites[ID].get(); 
+  return m_Sprites[ID].get();
 }
 
 Sprite* SpriteBase::Get(size_t ID)
